feat(frontend)!: Added applicationTrackerApi with token for reuse
Added a service folder with an api.ts with the applicationTrackerApi.
Exchanged old axios apis with the new API to get them up and running with token.

Fixed Issue with getting everything running in docker compose.

Note: to get everything running "authentication" needs to be added as domain loopback domain.
(On Linux: edit /etc/hosts "127.0.0.1 localhost authentication" instead of "127.0.0.1 localhost".)